Add a Basic "Skript Integration Fix" Toggle
This is kind of poorly made (new to java), so it probably misses
something and may cause some bugs. But the reason is the following: In
the SpectateListener, the general cancel on click events impedes Skript
to handle it, thus disallowing for custom menu items (like a leave
"button"). Thus, I first tried removing that single cancelation, and it
actually worked, but I wanted to make it an optional toggle, and commit
it to the original plugin, so more people can enjoy this plugin in a
more customized way :)

As mentioned in the yml option, this option should only be set to true
if you're going to use Skript, or another plugin that will cancel and
handle the click event. Otherwise, you'd allow people to interact with
the world...